j ai ajoute la fonction bridge_into_first qui bridge et remplace first en appelant free : il doit etre malloc, l input est maintenant aussi concatene en une seule string en prenant en compte les sauts de ligne, ce seras utile pour l historique, j'ai mis la fonction print_token dans un fichier temp.c pour se souvenir de la supprimer
