+++ config/sources.mk
@@ -3,10 +3,10 @@
 #                                                         :::      ::::::::    #
 #    sources.mk                                         :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: root <root@student.42.fr>                  +#+  +:+       +#+         #
+#    By: brappo <brappo@student.42.fr>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/03/31 20:47:09 by mhotting          #+#    #+#              #
-#    Updated: 2024/04/02 20:53:56 by root             ###   ########.fr        #
+#    Updated: 2024/04/03 12:20:50 by brappo           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -67,7 +67,8 @@ TOKENR_FILES			=	handle_quote.c				\
 							token_recognition_utils.c	\
 							token_recognition.c			\
 							tokenize_str.c				\
-							token_list_merge.c
+							token_list_merge.c			\
+							temp.c
 TOKENR					=	$(addprefix $(TOKENR_DIR), $(TOKENR_FILES))
 
 # PROMPT
