+++ srcs/main/minishell.c
@@ -6,7 +6,7 @@
 /*   By: brappo <brappo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/27 10:14:16 by mhotting          #+#    #+#             */
-/*   Updated: 2024/04/03 11:07:51 by brappo           ###   ########.fr       */
+/*   Updated: 2024/04/03 12:22:26 by brappo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,6 +31,7 @@ int	main(void)
 			break ;
 		}
 		token_recognition(&shell);
+		printf("input : \n%s\n", shell.input);
 		ft_lstprint(shell.tokens, print_token);
 		ft_lstclear(&shell.tokens, t_token_free);
 		free(shell.input);
