+++ srcs/tokenRecognition/token_list_merge.c
@@ -6,7 +6,7 @@
 /*   By: brappo <brappo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/02 20:42:20 by root              #+#    #+#             */
-/*   Updated: 2024/04/03 11:32:57 by brappo           ###   ########.fr       */
+/*   Updated: 2024/04/03 12:17:57 by brappo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
