+++ srcs/tokenRecognition/token_recognition_utils.c
@@ -6,7 +6,7 @@
 /*   By: brappo <brappo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/27 15:09:19 by brappo            #+#    #+#             */
-/*   Updated: 2024/04/03 10:10:32 by brappo           ###   ########.fr       */
+/*   Updated: 2024/04/03 12:20:28 by brappo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,22 +47,7 @@ bool	is_blank(char character)
 	return (false);
 }
 
-void	print_token(void *token_void)
+void	token_error(t_minishell *shell)
 {
-	t_token	*token;
-
-	if (token_void == NULL)
-		return ;
-	token = (t_token *)token_void;
-	if (token->str != NULL)
-		printf("%s : ", token->str);
-	printf("\033[0;35m");
-	if (token->type == END)
-		printf("END");
-	else if (token->type == WORD)
-		printf("WORD");
-	else if (token->type == OPERATOR)
-		printf("OPERATOR");
-	printf("\033[0m");
-	printf("\n");
+	handle_error(shell, TOKENIZATION_ERROR, EXIT_FAILURE);
 }
